various cleanup/bugfixes
