+++ src/plugins/intellisense/view.less
@@ -61,11 +61,6 @@ div:focus #intellisense-container {
   border-top: 1px solid #bbbbbb;
 }
 
-span[data-is-comma="true"]::after {
-  content: " ";
-  display: block;
-}
-
 .jump-to-def-btn {
   border-radius: 5px;
   padding: 2px;
