+++ src/plugins/pillbox-menus/components/Menu.tsx
@@ -24,7 +24,13 @@ export function MenuFunc(controller: PillboxMenus) {
 }
 
 const categoryPlugins: Record<string, PluginID[]> = {
-  core: ["builtin-settings", "GLesmos", "video-creator", "text-mode"],
+  core: [
+    "builtin-settings",
+    "GLesmos",
+    "video-creator",
+    "text-mode",
+    "intellisense",
+  ],
   utility: [
     "wolfram2desmos",
     "pin-expressions",
@@ -34,7 +40,6 @@ const categoryPlugins: Record<string, PluginID[]> = {
     "duplicate-expression-hotkey",
     "shift-enter-newline",
     "folder-tools",
-    "intellisense",
   ],
   visual: [
     "set-primary-color",
